some refactoring
update dependencies version
