+++ requirements-dev.txt
@@ -1,4 +1,4 @@
 -r requirements.txt
-flake8==3.8.4
+flake8==3.9.0
 jupyterlab==3.0.12
-pytest==6.2.1
+pytest==6.2.2
