+++ requirements.txt
@@ -1,14 +1,14 @@
-flair==0.7
-notebook==6.1.6
+flair==0.8.0.post1
+notebook==6.3.0
 opencv-contrib-python==4.5.1.48
-openpyxl==3.0.5
-pandas==1.1.5
+openpyxl==3.0.7
+pandas==1.2.3
 pdfminer.six==20201018
-prettytable==2.0.0
-PyMuPDF==1.18.6
-phonenumbers==8.12.17
+prettytable==2.1.0
+PyMuPDF==1.18.10
+phonenumbers==8.12.20
 seaborn==0.11.1
-sentencepiece==0.1.91
+sentencepiece==0.1.95
 spacy==2.3.5
 stanza==1.1.1
 styleframe==3.0.6
